Employee: Added support to barcode scanner
* Also changed viewProductFromCatalog to accept Long as barcode instead of Int.

Related issues:

Add support to barcode scanner #192
Change viewProductFromCatalog to accept Long as barcode instead of Int #193

+++ Employee/src/main/java/EmployeeContracts/IWorker.java
@@ -62,7 +62,7 @@ public interface IWorker {
 	 * @throws UnknownSenderID 
 	 * @throws InvalidParameter 
 	 */
-	CatalogProduct viewProductFromCatalog(int barcode) throws InvalidParameter, CriticalError, EmployeeNotConnected, ProductNotExistInCatalog;
+	CatalogProduct viewProductFromCatalog(long barcode) throws InvalidParameter, CriticalError, EmployeeNotConnected, ProductNotExistInCatalog;
 
 	/**
 	 * worker add product package to warehouse.
